requirements.txt
python-telegram-bot==20.7
pyrogram==2.0.106
telethon==1.28.5
tgcrypto==1.2.5
python-dotenv==1.0.0
pymongo==4.6.1
motor==3.3.2
cryptography==41.0.8
aiohttp==3.9.1
requests==2.31.0
python-multipart==0.0.6